split exceptions and functions into modules
