+++ src/schemata/exceptions.py
@@ -1,0 +1,13 @@
+class ConsentException(BaseException):
+    pass
+
+
+class ValidationError(BaseException):
+    pass
+
+
+ValidationErrors = (
+    ValidationError,
+    __import__("jsonschema").ValidationError,
+    ValueError,
+)
